refactor: PaymentProperties를 pay/dto/request → pay/config로 이동 (#331)
@ConfigurationProperties 설정 바인딩 클래스가 dto/request 패키지에 위치해
역할과 패키지 의도가 맞지 않아 pay/config로 이동.
참조하는 PaymentFeignConfig, PaymentAuthInterceptor import 경로 수정.
