+++ backend/widyu-api/src/main/java/com/widyu/pay/config/PaymentFeignConfig.java
@@ -1,6 +1,6 @@
 package com.widyu.pay.config;
 
-import com.widyu.pay.dto.request.PaymentProperties;
+import com.widyu.pay.config.PaymentProperties;
 import feign.Request;
 import java.util.concurrent.TimeUnit;
 import org.springframework.context.annotation.Bean;
